Calculate missing observation summary

Calculates the summary of missingness in a data set.

calcmissing(obj, ...)

# S3 method for data.frame
calcmissing(obj, MRO.case = FALSE, print = TRUE, final = TRUE, ...)

# S3 method for mro
calcmissing(obj, ...)

Arguments

obj
An object

...
additional arguments

MRO.case
does something with rownames

print
logical, should we print the thing?

final
logical, is this final?

Value

Missing value object

Methods (by class)

data.frame: Method for a dataframe

mro: accepts a whole mr.object , which is first mro.mat, second element labels, third element the input data frame.

See also

Author

Junjie Zeng

Examples

calcmissing(census.at.school.5000[,1:20])
#>          gender  age country_en country_mi height languages ethnicother_mi
#> numMiss      15   26         39         39    220       227            445
#> percMiss   0.3% 0.5%       0.8%       0.8%   4.4%      4.5%           8.9%
#>          Total
#> numMiss   5000
#> percMiss  100%
#> 
#>       gender age country_en country_mi height languages ethnicother_mi Freq
#>            1   1          1          1      1         1              1 4109
#>            1   1          1          1      1         1              0  390
#>            1   1          1          1      1         0              1  191
#>            1   1          1          1      0         1              1  183
#>            1   1          0          0      1         1              1   22
#>            1   0          1          1      1         1              1   21
#>            1   1          1          1      0         1              0   18
#>            1   1          1          1      1         0              0   16
#>            1   1          1          1      0         0              1   14
#>            1   1          0          0      1         1              0   13
#>            0   1          1          1      1         1              1   10
#>            0   1          1          1      1         1              0    3
#>            1   1          1          1      0         0              0    1
#>            1   1          0          0      1         0              1    1
#>            1   1          0          0      1         0              0    1
#>            1   1          0          0      0         0              0    1
#>            1   0          1          1      0         1              1    1
#>            1   0          1          1      1         0              1    1
#>            1   0          1          1      0         1              0    1
#>            1   0          0          0      1         0              0    1
#>            0   1          1          1      0         1              1    1
#>            0   0          1          1      1         1              1    1
#> Total     15  26         39         39    220       227            445 5000
#>       Percentage
#>             82.2
#>              7.8
#>              3.8
#>              3.7
#>              0.4
#>              0.4
#>              0.4
#>              0.3
#>              0.3
#>              0.3
#>              0.2
#>              0.1
#>              0.0
#>              0.0
#>              0.0
#>              0.0
#>              0.0
#>              0.0
#>              0.0
#>              0.0
#>              0.0
#>              0.0
#> Total      100.0