Calculate MRO inference

Calculates required proportions, their differences, variance-covariance matrices, standard errors of differences, and comparison intervals for differences, over all of the data. To compute values over various subsets of another explanatory variable, see by.

mroPara(obj, conf.levels = 1.96, nonparallel = NULL)

Arguments

obj
an MRO object created by iNZightMR

conf.levels
confidence level to use, default is 1.96 for 95% intervals

nonparallel
Should these things be parallel?

Value

An object of class mrocalc

See also

Examples

mr <- iNZightMR(online ~ onlinegame + onlinevideo + onlinemusic,
    data = census.at.school.5000)
mrp <- mroPara(mr)